+++ src/realm/sync/network/default_socket.cpp
@@ -470,8 +470,11 @@ void DefaultWebSocketImpl::initiate_websocket_handshake()
 ///
 
 DefaultSocketProvider::DefaultSocketProvider(const std::shared_ptr<util::Logger>& logger,
-                                             const std::string user_agent, AutoStart auto_start)
+                                             const std::string user_agent,
+                                             const std::shared_ptr<BindingCallbackThreadObserver>& observer_ptr,
+                                             AutoStart auto_start)
     : m_logger_ptr{logger}
+    , m_observer_ptr{observer_ptr}
     , m_service{}
     , m_random{}
     , m_user_agent{user_agent}
@@ -510,6 +513,7 @@ void DefaultSocketProvider::start()
 
     m_logger_ptr->trace("Default event loop: start()");
     REALM_ASSERT(m_state == State::Stopped);
+
     do_state_update(lock, State::Starting);
     m_thread = std::thread{&DefaultSocketProvider::event_loop, this};
     // Wait for the thread to start before continuing
@@ -519,10 +523,11 @@ void DefaultSocketProvider::start()
 void DefaultSocketProvider::event_loop()
 {
     m_logger_ptr->trace("Default event loop: thread running");
+    // Calls will_destroy_thread() when destroyed
     auto will_destroy_thread = util::make_scope_exit([&]() noexcept {
         m_logger_ptr->trace("Default event loop: thread exiting");
-        if (g_binding_callback_thread_observer)
-            g_binding_callback_thread_observer->will_destroy_thread();
+        if (m_observer_ptr)
+            m_observer_ptr->will_destroy_thread();
 
         std::unique_lock<std::mutex> lock(m_mutex);
         // Did we get here due to an unhandled exception?
@@ -533,8 +538,8 @@ void DefaultSocketProvider::event_loop()
         std::notify_all_at_thread_exit(m_state_cv, std::move(lock));
     });
 
-    if (g_binding_callback_thread_observer)
-        g_binding_callback_thread_observer->did_create_thread();
+    if (m_observer_ptr)
+        m_observer_ptr->did_create_thread();
 
     {
         std::lock_guard<std::mutex> lock(m_mutex);
@@ -572,9 +577,8 @@ void DefaultSocketProvider::event_loop()
         do_state_update(lock, State::Stopping);
         lock.unlock();
         m_logger_ptr->error("Default event loop exception: ", e.what());
-        if (g_binding_callback_thread_observer)
-            g_binding_callback_thread_observer->handle_error(e);
-        else
+        // If the observer_ptr is not set or the error was not handled, then throw the exception
+        if (!m_observer_ptr || !m_observer_ptr->handle_error(e))
             throw;
     }
 }